+++ src/ast/compare.rs
@@ -1,45 +1,45 @@
-use crate::ast::{Column, ConditionTree, DatabaseValue, Expression};
+use crate::ast::{Column, ConditionTree, Expression};
 use std::borrow::Cow;
 
 /// For modeling comparison expression
 #[derive(Debug, Clone, PartialEq)]
 pub enum Compare<'a> {
     /// `left = right`
-    Equals(Box<DatabaseValue<'a>>, Box<DatabaseValue<'a>>),
+    Equals(Box<Expression<'a>>, Box<Expression<'a>>),
     /// `left <> right`
-    NotEquals(Box<DatabaseValue<'a>>, Box<DatabaseValue<'a>>),
+    NotEquals(Box<Expression<'a>>, Box<Expression<'a>>),
     /// `left < right`
-    LessThan(Box<DatabaseValue<'a>>, Box<DatabaseValue<'a>>),
+    LessThan(Box<Expression<'a>>, Box<Expression<'a>>),
     /// `left <= right`
-    LessThanOrEquals(Box<DatabaseValue<'a>>, Box<DatabaseValue<'a>>),
+    LessThanOrEquals(Box<Expression<'a>>, Box<Expression<'a>>),
     /// `left > right`
-    GreaterThan(Box<DatabaseValue<'a>>, Box<DatabaseValue<'a>>),
+    GreaterThan(Box<Expression<'a>>, Box<Expression<'a>>),
     /// `left >= right`
-    GreaterThanOrEquals(Box<DatabaseValue<'a>>, Box<DatabaseValue<'a>>),
+    GreaterThanOrEquals(Box<Expression<'a>>, Box<Expression<'a>>),
     /// `left IN (..)`
-    In(Box<DatabaseValue<'a>>, Box<DatabaseValue<'a>>),
+    In(Box<Expression<'a>>, Box<Expression<'a>>),
     /// `left NOT IN (..)`
-    NotIn(Box<DatabaseValue<'a>>, Box<DatabaseValue<'a>>),
+    NotIn(Box<Expression<'a>>, Box<Expression<'a>>),
     /// `left LIKE %..%`
-    Like(Box<DatabaseValue<'a>>, Cow<'a, str>),
+    Like(Box<Expression<'a>>, Cow<'a, str>),
     /// `left NOT LIKE %..%`
-    NotLike(Box<DatabaseValue<'a>>, Cow<'a, str>),
+    NotLike(Box<Expression<'a>>, Cow<'a, str>),
     /// `left LIKE ..%`
-    BeginsWith(Box<DatabaseValue<'a>>, Cow<'a, str>),
+    BeginsWith(Box<Expression<'a>>, Cow<'a, str>),
     /// `left NOT LIKE ..%`
-    NotBeginsWith(Box<DatabaseValue<'a>>, Cow<'a, str>),
+    NotBeginsWith(Box<Expression<'a>>, Cow<'a, str>),
     /// `left LIKE %..`
-    EndsInto(Box<DatabaseValue<'a>>, Cow<'a, str>),
+    EndsInto(Box<Expression<'a>>, Cow<'a, str>),
     /// `left NOT LIKE %..`
-    NotEndsInto(Box<DatabaseValue<'a>>, Cow<'a, str>),
+    NotEndsInto(Box<Expression<'a>>, Cow<'a, str>),
     /// `value IS NULL`
-    Null(Box<DatabaseValue<'a>>),
+    Null(Box<Expression<'a>>),
     /// `value IS NOT NULL`
-    NotNull(Box<DatabaseValue<'a>>),
+    NotNull(Box<Expression<'a>>),
     /// `value` BETWEEN `left` AND `right`
-    Between(Box<DatabaseValue<'a>>, Box<DatabaseValue<'a>>, Box<DatabaseValue<'a>>),
+    Between(Box<Expression<'a>>, Box<Expression<'a>>, Box<Expression<'a>>),
     /// `value` NOT BETWEEN `left` AND `right`
-    NotBetween(Box<DatabaseValue<'a>>, Box<DatabaseValue<'a>>, Box<DatabaseValue<'a>>),
+    NotBetween(Box<Expression<'a>>, Box<Expression<'a>>, Box<Expression<'a>>),
 }
 
 impl<'a> From<Compare<'a>> for ConditionTree<'a> {
@@ -74,7 +74,7 @@ pub trait Comparable<'a> {
     /// ```
     fn equals<T>(self, comparison: T) -> Compare<'a>
     where
-        T: Into<DatabaseValue<'a>>;
+        T: Into<Expression<'a>>;
 
     /// Tests if both sides are not the same value.
     ///
@@ -94,7 +94,7 @@ pub trait Comparable<'a> {
     /// ```
     fn not_equals<T>(self, comparison: T) -> Compare<'a>
     where
-        T: Into<DatabaseValue<'a>>;
+        T: Into<Expression<'a>>;
 
     /// Tests if the left side is smaller than the right side.
     ///
@@ -114,7 +114,7 @@ pub trait Comparable<'a> {
     /// ```
     fn less_than<T>(self, comparison: T) -> Compare<'a>
     where
-        T: Into<DatabaseValue<'a>>;
+        T: Into<Expression<'a>>;
 
     /// Tests if the left side is smaller than the right side or the same.
     ///
@@ -134,7 +134,7 @@ pub trait Comparable<'a> {
     /// ```
     fn less_than_or_equals<T>(self, comparison: T) -> Compare<'a>
     where
-        T: Into<DatabaseValue<'a>>;
+        T: Into<Expression<'a>>;
 
     /// Tests if the left side is bigger than the right side.
     ///
@@ -154,7 +154,7 @@ pub trait Comparable<'a> {
     /// ```
     fn greater_than<T>(self, comparison: T) -> Compare<'a>
     where
-        T: Into<DatabaseValue<'a>>;
+        T: Into<Expression<'a>>;
 
     /// Tests if the left side is bigger than the right side or the same.
     ///
@@ -174,7 +174,7 @@ pub trait Comparable<'a> {
     /// ```
     fn greater_than_or_equals<T>(self, comparison: T) -> Compare<'a>
     where
-        T: Into<DatabaseValue<'a>>;
+        T: Into<Expression<'a>>;
 
     /// Tests if the left side is included in the right side collection.
     ///
@@ -191,7 +191,7 @@ pub trait Comparable<'a> {
     /// ```
     fn in_selection<T>(self, selection: T) -> Compare<'a>
     where
-        T: Into<DatabaseValue<'a>>;
+        T: Into<Expression<'a>>;
 
     /// Tests if the left side is not included in the right side collection.
     ///
@@ -209,7 +209,7 @@ pub trait Comparable<'a> {
     /// ```
     fn not_in_selection<T>(self, selection: T) -> Compare<'a>
     where
-        T: Into<DatabaseValue<'a>>;
+        T: Into<Expression<'a>>;
 
     /// Tests if the left side includes the right side string.
     ///
@@ -369,8 +369,8 @@ pub trait Comparable<'a> {
     /// ```
     fn between<T, V>(self, left: T, right: V) -> Compare<'a>
     where
-        T: Into<DatabaseValue<'a>>,
-        V: Into<DatabaseValue<'a>>;
+        T: Into<Expression<'a>>,
+        V: Into<Expression<'a>>;
 
     /// Tests if the value is not between two given values.
     ///
@@ -388,8 +388,8 @@ pub trait Comparable<'a> {
     /// ```
     fn not_between<T, V>(self, left: T, right: V) -> Compare<'a>
     where
-        T: Into<DatabaseValue<'a>>,
-        V: Into<DatabaseValue<'a>>;
+        T: Into<Expression<'a>>,
+        V: Into<Expression<'a>>;
 }
 
 impl<'a, U> Comparable<'a> for U
@@ -398,74 +398,74 @@ where
 {
     fn equals<T>(self, comparison: T) -> Compare<'a>
     where
-        T: Into<DatabaseValue<'a>>,
+        T: Into<Expression<'a>>,
     {
         let col: Column<'a> = self.into();
-        let val: DatabaseValue<'a> = col.into();
+        let val: Expression<'a> = col.into();
 
         val.equals(comparison)
     }
 
     fn not_equals<T>(self, comparison: T) -> Compare<'a>
     where
-        T: Into<DatabaseValue<'a>>,
+        T: Into<Expression<'a>>,
     {
         let col: Column<'a> = self.into();
-        let val: DatabaseValue<'a> = col.into();
+        let val: Expression<'a> = col.into();
         val.not_equals(comparison)
     }
 
     fn less_than<T>(self, comparison: T) -> Compare<'a>
     where
-        T: Into<DatabaseValue<'a>>,
+        T: Into<Expression<'a>>,
     {
         let col: Column<'a> = self.into();
-        let val: DatabaseValue<'a> = col.into();
+        let val: Expression<'a> = col.into();
         val.less_than(comparison)
     }
 
     fn less_than_or_equals<T>(self, comparison: T) -> Compare<'a>
     where
-        T: Into<DatabaseValue<'a>>,
+        T: Into<Expression<'a>>,
     {
         let col: Column<'a> = self.into();
-        let val: DatabaseValue<'a> = col.into();
+        let val: Expression<'a> = col.into();
         val.less_than_or_equals(comparison)
     }
 
     fn greater_than<T>(self, comparison: T) -> Compare<'a>
     where
-        T: Into<DatabaseValue<'a>>,
+        T: Into<Expression<'a>>,
     {
         let col: Column<'a> = self.into();
-        let val: DatabaseValue<'a> = col.into();
+        let val: Expression<'a> = col.into();
         val.greater_than(comparison)
     }
 
     fn greater_than_or_equals<T>(self, comparison: T) -> Compare<'a>
     where
-        T: Into<DatabaseValue<'a>>,
+        T: Into<Expression<'a>>,
     {
         let col: Column<'a> = self.into();
-        let val: DatabaseValue<'a> = col.into();
+        let val: Expression<'a> = col.into();
         val.greater_than_or_equals(comparison)
     }
 
     fn in_selection<T>(self, selection: T) -> Compare<'a>
     where
-        T: Into<DatabaseValue<'a>>,
+        T: Into<Expression<'a>>,
     {
         let col: Column<'a> = self.into();
-        let val: DatabaseValue<'a> = col.into();
+        let val: Expression<'a> = col.into();
         val.in_selection(selection)
     }
 
     fn not_in_selection<T>(self, selection: T) -> Compare<'a>
     where
-        T: Into<DatabaseValue<'a>>,
+        T: Into<Expression<'a>>,
     {
         let col: Column<'a> = self.into();
-        let val: DatabaseValue<'a> = col.into();
+        let val: Expression<'a> = col.into();
         val.not_in_selection(selection)
     }
 
@@ -474,7 +474,7 @@ where
         T: Into<Cow<'a, str>>,
     {
         let col: Column<'a> = self.into();
-        let val: DatabaseValue<'a> = col.into();
+        let val: Expression<'a> = col.into();
         val.like(pattern)
     }
 
@@ -483,7 +483,7 @@ where
         T: Into<Cow<'a, str>>,
     {
         let col: Column<'a> = self.into();
-        let val: DatabaseValue<'a> = col.into();
+        let val: Expression<'a> = col.into();
         val.not_like(pattern)
     }
 
@@ -492,7 +492,7 @@ where
         T: Into<Cow<'a, str>>,
     {
         let col: Column<'a> = self.into();
-        let val: DatabaseValue<'a> = col.into();
+        let val: Expression<'a> = col.into();
         val.begins_with(pattern)
     }
 
@@ -501,7 +501,7 @@ where
         T: Into<Cow<'a, str>>,
     {
         let col: Column<'a> = self.into();
-        let val: DatabaseValue<'a> = col.into();
+        let val: Expression<'a> = col.into();
         val.not_begins_with(pattern)
     }
 
@@ -510,7 +510,7 @@ where
         T: Into<Cow<'a, str>>,
     {
         let col: Column<'a> = self.into();
-        let val: DatabaseValue<'a> = col.into();
+        let val: Expression<'a> = col.into();
         val.ends_into(pattern)
     }
 
@@ -519,39 +519,39 @@ where
         T: Into<Cow<'a, str>>,
     {
         let col: Column<'a> = self.into();
-        let val: DatabaseValue<'a> = col.into();
+        let val: Expression<'a> = col.into();
         val.not_ends_into(pattern)
     }
 
     fn is_null(self) -> Compare<'a> {
         let col: Column<'a> = self.into();
-        let val: DatabaseValue<'a> = col.into();
+        let val: Expression<'a> = col.into();
         val.is_null()
     }
 
     fn is_not_null(self) -> Compare<'a> {
         let col: Column<'a> = self.into();
-        let val: DatabaseValue<'a> = col.into();
+        let val: Expression<'a> = col.into();
         val.is_not_null()
     }
 
     fn between<T, V>(self, left: T, right: V) -> Compare<'a>
     where
-        T: Into<DatabaseValue<'a>>,
-        V: Into<DatabaseValue<'a>>,
+        T: Into<Expression<'a>>,
+        V: Into<Expression<'a>>,
     {
         let col: Column<'a> = self.into();
-        let val: DatabaseValue<'a> = col.into();
+        let val: Expression<'a> = col.into();
         val.between(left, right)
     }
 
     fn not_between<T, V>(self, left: T, right: V) -> Compare<'a>
     where
-        T: Into<DatabaseValue<'a>>,
-        V: Into<DatabaseValue<'a>>,
+        T: Into<Expression<'a>>,
+        V: Into<Expression<'a>>,
     {
         let col: Column<'a> = self.into();
-        let val: DatabaseValue<'a> = col.into();
+        let val: Expression<'a> = col.into();
         val.not_between(left, right)
     }
 }