+++ src/ast/function/count.rs
@@ -1,10 +1,10 @@
 use super::Function;
-use crate::ast::DatabaseValue;
+use crate::ast::Expression;
 
 #[derive(Debug, Clone, PartialEq)]
 /// Returns the number of rows that matches a specified criteria.
 pub struct Count<'a> {
-    pub(crate) exprs: Vec<DatabaseValue<'a>>,
+    pub(crate) exprs: Vec<Expression<'a>>,
 }
 
 /// Count of the underlying table where the given expression is not null.
@@ -17,7 +17,7 @@ pub struct Count<'a> {
 /// ```
 pub fn count<'a, T>(expr: T) -> Function<'a>
 where
-    T: Into<DatabaseValue<'a>>,
+    T: Into<Expression<'a>>,
 {
     let fun = Count {
         exprs: vec![expr.into()],