+++ CHANGELOG.md
@@ -105,7 +105,7 @@ Breaking changes ahead
 ## v0.1.6
 
 - Loosen up certain vector-taking functions to use `IntoIter`
-- Fix the `columns` method to actually use `Column` instead of a plain `DatabaseValue`
+- Fix the `columns` method to actually use `Column` instead of a plain `Expression`
 
 ## v0.1.5